【イベント】【間取り編集】ピクセルを消せるようにする #47  ツールに関する処理のディレクトリを変更&共通部分の切り出し
